move accessor to its own module
