+++ xarray/plot/dataarray_plot.py
@@ -2178,775 +2178,3 @@ def surface(x, y, z, ax, **kwargs) -> Poly3DCollection:
     """
     primitive = ax.plot_surface(x, y, z, **kwargs)
     return primitive
-
-
-class DataArrayPlotAccessor:
-    """
-    Enables use of xarray.plot functions as attributes on a DataArray.
-    For example, DataArray.plot.imshow
-    """
-
-    _da: DataArray
-
-    __slots__ = ("_da",)
-    __doc__ = plot.__doc__
-
-    def __init__(self, darray: DataArray) -> None:
-        self._da = darray
-
-    # Should return Any such that the user does not run into problems
-    # with the many possible return values
-    @functools.wraps(plot, assigned=("__doc__", "__annotations__"))
-    def __call__(self, **kwargs) -> Any:
-        return plot(self._da, **kwargs)
-
-    @functools.wraps(hist)
-    def hist(self, *args, **kwargs) -> tuple[np.ndarray, np.ndarray, BarContainer]:
-        return hist(self._da, *args, **kwargs)
-
-    @overload
-    def line(
-        self,
-        *args: Any,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        figsize: Iterable[float] | None = None,
-        aspect: MPLAspectOptions = None,
-        size: float | None = None,
-        ax: Axes | None = None,
-        hue: Hashable | None = None,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        xincrease: bool | None = None,
-        yincrease: bool | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        add_legend: bool = True,
-        _labels: bool = True,
-        **kwargs: Any,
-    ) -> list[Line3D]:
-        ...
-
-    @overload
-    def line(
-        self,
-        *args: Any,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        aspect: MPLAspectOptions = None,
-        size: float | None = None,
-        ax: Axes | None = None,
-        hue: Hashable | None = None,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        xincrease: bool | None = None,
-        yincrease: bool | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        add_legend: bool = True,
-        _labels: bool = True,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def line(
-        self,
-        *args: Any,
-        row: Hashable | None = None,
-        col: Hashable,  # wrap -> FacetGrid
-        figsize: Iterable[float] | None = None,
-        aspect: MPLAspectOptions = None,
-        size: float | None = None,
-        ax: Axes | None = None,
-        hue: Hashable | None = None,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        xincrease: bool | None = None,
-        yincrease: bool | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        add_legend: bool = True,
-        _labels: bool = True,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(line)
-    def line(self, *args, **kwargs) -> list[Line3D] | FacetGrid:
-        return line(self._da, *args, **kwargs)
-
-    @overload
-    def step(
-        self,
-        *args: Any,
-        where: Literal["pre", "post", "mid"] = "pre",
-        drawstyle: str | None = None,
-        ds: str | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        **kwargs: Any,
-    ) -> list[Line3D]:
-        ...
-
-    @overload
-    def step(
-        self,
-        *args: Any,
-        where: Literal["pre", "post", "mid"] = "pre",
-        drawstyle: str | None = None,
-        ds: str | None = None,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def step(
-        self,
-        *args: Any,
-        where: Literal["pre", "post", "mid"] = "pre",
-        drawstyle: str | None = None,
-        ds: str | None = None,
-        row: Hashable | None = None,
-        col: Hashable,  # wrap -> FacetGrid
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(step)
-    def step(self, *args, **kwargs) -> list[Line3D] | FacetGrid:
-        return step(self._da, *args, **kwargs)
-
-    @functools.wraps(scatter)
-    def _scatter(self, *args, **kwargs):
-        return scatter(self._da, *args, **kwargs)
-
-    @overload
-    def imshow(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> AxesImage:
-        ...
-
-    @overload
-    def imshow(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: Hashable | None = None,
-        *,
-        col: Hashable,  # wrap -> FacetGrid
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def imshow(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        *,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(imshow)
-    def imshow(self, *args, **kwargs) -> AxesImage:
-        return imshow(self._da, *args, **kwargs)
-
-    @overload
-    def contour(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> QuadContourSet:
-        ...
-
-    @overload
-    def contour(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: Hashable | None = None,
-        *,
-        col: Hashable,  # wrap -> FacetGrid
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def contour(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        *,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(contour)
-    def contour(self, *args, **kwargs) -> QuadContourSet:
-        return contour(self._da, *args, **kwargs)
-
-    @overload
-    def contourf(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> QuadContourSet:
-        ...
-
-    @overload
-    def contourf(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: Hashable | None = None,
-        *,
-        col: Hashable,  # wrap -> FacetGrid
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def contourf(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        *,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(contourf)
-    def contourf(self, *args, **kwargs) -> QuadContourSet:
-        return contourf(self._da, *args, **kwargs)
-
-    @overload
-    def pcolormesh(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> QuadMesh:
-        ...
-
-    @overload
-    def pcolormesh(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: Hashable | None = None,
-        *,
-        col: Hashable,  # wrap -> FacetGrid
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def pcolormesh(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        *,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(pcolormesh)
-    def pcolormesh(self, *args, **kwargs) -> QuadMesh:
-        return pcolormesh(self._da, *args, **kwargs)
-
-    @overload
-    def surface(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: None = None,  # no wrap -> primitive
-        col: None = None,  # no wrap -> primitive
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> Poly3DCollection:
-        ...
-
-    @overload
-    def surface(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        row: Hashable | None = None,
-        *,
-        col: Hashable,  # wrap -> FacetGrid
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @overload
-    def surface(
-        self,
-        x: Hashable | None = None,
-        y: Hashable | None = None,
-        figsize: Iterable[float] | None = None,
-        size: float | None = None,
-        aspect: MPLAspectOptions = None,
-        ax: Axes | None = None,
-        *,
-        row: Hashable,  # wrap -> FacetGrid
-        col: Hashable | None = None,
-        col_wrap: int | None = None,
-        xincrease: bool | None = True,
-        yincrease: bool | None = True,
-        add_colorbar: bool | None = None,
-        add_labels: bool = True,
-        vmin: float | None = None,
-        vmax: float | None = None,
-        cmap=None,
-        center=None,
-        robust: bool = False,
-        extend=None,
-        levels=None,
-        infer_intervals=None,
-        colors=None,
-        subplot_kws: dict[str, Any] | None = None,
-        cbar_ax: Axes | None = None,
-        cbar_kwargs: dict[str, Any] | None = None,
-        xscale: MPLScaleOptions = None,
-        yscale: MPLScaleOptions = None,
-        xticks: ArrayLike | None = None,
-        yticks: ArrayLike | None = None,
-        xlim: ArrayLike | None = None,
-        ylim: ArrayLike | None = None,
-        norm: Normalize | None = None,
-        **kwargs: Any,
-    ) -> FacetGrid:
-        ...
-
-    @functools.wraps(surface)
-    def surface(self, *args, **kwargs) -> Poly3DCollection:
-        return surface(self._da, *args, **kwargs)
